requirements.txt
absl-py==1.0.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
cvzone==1.5.4
cycler==0.11.0
dataclasses-json==0.6.1
distlib==0.3.3
Django==3.2.8
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
filelock==3.3.1
frozenlist==1.4.0
fsspec==2023.9.2
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httplib2==0.21.0
httptools==0.2.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.3.2
langchain==0.0.304
langsmith==0.0.41
llama-index==0.8.37
lxml==4.9.1
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.4.3
mediapipe==0.8.9.1
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
nest-asyncio==1.5.8
nibabel==3.2.2
nltk==3.8.1
numexpr==2.8.7
numpy==1.21.2
openai==0.28.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
orjson==3.6.5
packaging==21.3
pandas==1.3.4
parsel==1.6.0
pbr==5.6.0
Pillow==8.4.0
platformdirs==2.4.0
priority==1.3.0
Protego==0.1.16
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycipher==0.5.2
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.8.2
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
queuelib==1.6.2
regex==2023.8.8
requests==2.26.0
rsa==4.9
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
spotipy==2.19.0
SQLAlchemy==2.0.21
sqlparse==0.4.2
starlette==0.16.0
stevedore==3.5.0
tenacity==8.2.3
tiktoken==0.5.1
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.66.1
tree-sitter==0.20.2
tree-sitter-languages==1.7.0
turtle==0.0.1
Twisted==21.7.0
typing-inspect==0.9.0
typing_extensions==4.8.0
ujson==4.3.0
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
w3lib==1.22.0
watchgod==0.7
websockets==10.1
wordcloud==1.8.2.2
yarl==1.9.2
zope.interface==5.4.0
